vaccination implementation

+++ src/main/java/org/springframework/samples/petclinic/vaccine/predicate/PetPredicates.java
@@ -10,7 +10,8 @@ public class PetPredicates {
 		return pet -> Boolean.TRUE.equals(pet.getActive());
 	}
 
-	public static Predicate<Pet> hasId(Long petId) {
+	public static Predicate<Pet> hasId(Integer petId) {
 		return pet -> petId.equals(pet.getId());
 	}
+
 }
